requirements.txt
# Banco de dados
psycopg[binary]>=3.2

# Gerador de PDF
fpdf2==2.7.9

# Seguranca
bcrypt==4.1.2

# Variaveis de Ambiente
python-dotenv==1.0.0